requirements.txt
python-dotenv==1.0.0
psutil==5.9.8
aiogram==2.24